Exclude final field from classes and interfaced when ordering fields. Use long to do bit operations.

+++ src/ch/unibe/iam/scg/ContextInfo.java
@@ -4,5 +4,5 @@ public class ContextInfo {
 	public boolean global= false;
 	public Object next = null;
 	public Object prev = null;
-	public int dirty = 0;
+	public long dirty = 0;
 }
